chore(api): add logic to api client
