+++ crates/api/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod http;
+
 use http::HttpClient;
 use std::sync::LazyLock;
 use tokio::sync::RwLock;
